requirements.txt
streamlit==1.39.0
pandas==2.2.2
requests==2.32.3
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
python-docx==1.1.2
openai==1.50.2
PyYAML==6.0.2
plotly==5.24.1
